added 16 bit image support to coloc threshold plugin
enabled proper drawing of RGB colcoalisation map image/stack for 16 bit input data
but scaling the pixels to 8 bit range, same scaling for both chanels,
so there is no overflow when aking the rbg coloc map image.
Moved thie making of the coloc map image into its own method.
Used image statistic methos from imageJ to get the max pixel values of the 16 bit images
and put in a test to see of the images need scaling to 8 bit range or not.