Coloc 2: make the imageJ tool bar status message show the real values
for the 2Dhistograms and other result images displayed,
need to show the real value,
even if the image displayed is the log values of the result image,

Also need to show the bin ranges for each pixel in x and y of a 2D histogram
not just the position of the x and y in the 3D histogrtam image.

But , you can see the x y coords of the image by pressing Alt,
as per normal imageJ image windows when you dont want to see the calibrated x y coords.