Coloc 2: Simplify and reduce Pearsons interface and calculations
The interface offers now only one method for classic and fast implementation.
TwinValueRangeCursors are now used in both implementations to iterate
over the image data. With them one can have one implementation regardless
if one wants thresholded data or not.

AutoThresholdRegression now uses that and an own TwinValueRangeCursor to
iterate over the images. Due to that interface chagne it uses now the
Pearsons implementation choosen in the provided object in the constructor.
It reuses the cursor all the time and resets the thresholds of the predicats.

Furthermore a bug has been fixed where the stored Pearson R value was over-
written. It was caused by the call to Pearsons calculation during the
Costes significance test.