Coloc 2: Let the mask cursor repeat the mask if it went through it
A mask is usually a 2D image and to apply the MaskCursor easily to a stack
this mask should be repeated once it is done with it. That allows us to
specifiy only a 2D image as mask for a whole stack and basically act like
a RegionOfInterestCursor, but for based on that image.