Coloc 2: Add basic irregular ROI/mask test. Mask self with self and inverted
This test generates a random black/white noise image and uses first itself
and then an inverted version of it as mask. While iterating over it, the
pixel values are multiplied by each other. In the end the the product would
be one in the non-inverted case and zero in the other. This way one makes
sure, that. only ones and zeros are present in the result.