Coloc 2: Make Costes more robust, be more fault tolerant
In Pearson's can occur numerical problems, e.g. if there are too few pixels
or the images values sum up in a unlucky way (very seldom). If such a thing
happens, Costes will allow a specified number of retries (3) in randomisations.
If that still leads to an error, the exception is re-thrown.