Coloc 2: Add two more regular ROI test cases
The first one counts the pixels while walking over the ROI and asserts it
to be the same as the expected volume of the ROI box. The second one
makes sure the RoiImages offset and size is calculated correctly by
comparing it to the expected data for a particular test image.