Coloc 2: replace spaces in dialog items with underscores
In GenericDialogs, the labels are not only used to display to the
user but also to record macros. For the macros, however, only the part
until the first space is used, and to avoid ambiguity the GenericDialog
supports adding underscores instead of spaces (which will be replaced
before showing to the user, but are used in recorded macros).