Coloc 2: Allow masks with lower dimensionality and in there same extent
If an irregular ROI is used with a stack, the mask information will be
of a different dimensionality (e.g. 2) than the stack (e.g. 3). Therefore,
a valid mask of same dimensionality and extend as the image needs to be
projected into the image space. A method for that has been added and is
used for irregular ROIs and image masks. For now only masks of the same
extent as the image in the corresponding dimensions are allowed. The plugin
will abort if that constraint is not met.