Coloc 2: Indicate what kind of mask is used in a cleaner way
Internally ROIs and masks should be treated equally. So the interface
should not indicate greater differnces between them. Nevertheless, in some
cases it is good to know it the mask is regular (i.e. a  rectangle) or
irregular (e.g. a circle). In the Costes algorithm this information is
used to improve perfomance in one situation.