Coloc 2: fix test for creation of regular masks and make it simpler
Three bugs were in this test: the actual cursor forwarding was missing, the
ROI offset position was not inclueded and the wrong assert() was used.
These have been fixed. Moreover the test has been renamed to better
express what it is doing, some commens have been added and some lines
were rewritten in a simpler way.