Coloc 2: fix problem with masks having a z-offset that crashed Costes
If a mask is used, the Cosets significance test should shuffle only the
data within the mask and ignore the rest. If the mask started not at z=0,
but had an offset in z, the shuffleing failed and the plugin raised an
exception. This is fixed now. The Costes calculations did not use the
mask z-size, but the image z-size to determine how many blocks to generate.
With the wrong z-size it tried to create more blocks than possible.