Coloc 2: don't scale histogram intensities up if maximum < bins
The 2D histogram's abscissa shows the intensity values of channel
one and the ordinate the one of channel two. If the maximum
intensity is higher than the available number of bins in X and Y,
respectively, the calculated coodinates get scaled down so that
the maximum intensities are on the last bin available.

Up to now this also happened when the maximum intensity is lower
than the number of bins. Of course, the coodinates need up-scaling
in such a case. However, simple upscaling brings in artifacts of
empty bins. To get around this, the plugin now doesn't do up-scaling
anymore. The resulting histogram fills therefore only a part of
the image space, but doesn't show artifacts anymore.