Add test for commutativity of the auto thresholding regression
The added test makes sure it doesn't matter if one calculates the auto
threshold regression for two images A and B or between the same images,
but with image B in the first and image B in the second channel.

This tests fails currently and requires a fix in the threshold
regression code.