Auto threshold regression: waslk channel two if slope below -1
In 863fdbe it was introduced that the working threshold walks channel
one if the regression line leans more towards the abscissa (which
represents channel one). This test, however, only tested if the slope
was below 1 and didn't handle negative correlation correctly. Therefore,
this commit makes sure channel one is only used for the working
threshold if the regression line's slope is below 1 and above -1. This
makes sure channel two is used if the slope is below or equal to -1.
Only until this slope, the regression line leans towards the abscissa,
i.e. channel one. Channel two is now used if the slope falls below -1
(or is equal to it), because the projected distance will be larger for
channel two (because the regression line leans more towards the
ordinate).