Add original threshold regression of Costes
This is now the default, because it is very likely what people expect.
However, the current bisection method is still available as an option.
It can produce good thresholds if the images are suited for
colocalization analysis. The user has to choose what is used. The
selected method is also shown in the output.

Costes' regression starts at the maximum threshold and moves down the
regression line one by one until the Person colocalization between both
channels is zero. This is what was originally described in Costes'
publication. The bisection pattern, in turn, is usually faster, but
might produce different results then the sequential walk. This can
happen if the function of correlations of a specific threshold
combination is not monotonic.