input image stats/params/vals now in DataContainer
SingleWindowDisplay used to also add a bunch of nice image
stats/values/parameters gotten from imglib2. These were missing from PDF
output and the IJ.log output. So now all these nice stats are put in the
DataContainer as ValueResults, and used everywhere a ResultsHandler
implementation like SingleWindowDisplay or the PDFWriter handles the
value results. It means there are no longer separateresults and stats
sections... but at least the output is the same in every case.