use a job name with the mask hash for results naming
Make sure the results window, pdf file name and doc forst line, and
SingleWindowDisplay  title, and also the ValueResults list in the
DataContainer all have the same jobName text. This contains the two
input image names, the maskHash from imglib2 as a hopefully unique
identifier, so the extra file naming in the PDFWriter isnt needed, as
the mask hash will make it unique per mask, ROI,, run, and even with no
user selected mask or RIO... as imglib2 always seems to have one for us.

But why so much copy paste code in the three DataContainer methoods? Can
it be simplified? I had to write the same line three times... that cant
be good.