change autothreshold test to y-intercept:ch2Mean
ch2 max value might be a very high value outlier of noise. So using the
mean value is safer.

Also changed comments and warning message workings to add more detail,
and clarify that the y-intercept can be lower or higher than zero, and
that's bad in both cases. Now if the y-intercept is lower than zero, the
warning message still makes sense and doesn't says its too high when its
actually too low.