fix bug in Manders' coeff  pixel selection logic
manders coefficients were calculated correctly for the
ThresoldType.none  case - we see if the pixel is above thresold (zero)
in the OTHER channel .

However, for the .Aboive and .Below thresholdType cases, we were
checking if the pixel was above or below threshold in the SAME image
channel - that's wrong.

This bug probably had a small impact as the thresholds of the two
channels tend to be rather similar and quite low.  The results wont
change much in the majority of situations. Only when the thresholds are
very different will be bug be very obvious.

Also fix some typos and add comments so we can understand whats going on
where better.

Renamed some variables and a class to make them clearer and make sense
later when adding more split coefficients like fraction intensity and
pixels colocalised (%)