real noisy Manders test  - progress, but still broken
sorted out the generic types stuff and got rid of all the warnings in
eclipse, it compiles, but i still get a cant cast error at runtime. So
still broken.

fixed method in TestImageAcessor to open the mask image as Img so can
get its cursor for use in the TwinCursor

Fixed setting the thresholds manually by making new Type instances for
them. of the required UnsignedByteType.