Results: issue #33, channel name in image name for results.
put the Ch1_ or Ch2_ text prefic in front of the image channel name
to the user know which is which.

I notice again the method in question has three versions,
which share much code, and should be thus refactored to
prevent copy paste bugs... i had to do the same things 3x here.

Signed-off-by: daniel j. white <dan@chalkie.org.uk>