Fix Manders' split coefficients wrong channel check bug
Manders' split coefficients were calculated correctly for the
ThresoldType.none  case - we see if the pixel is above thresold (zero)
in the OTHER channel .

However, for the .Aboive and .Below thresholdType cases, we were
checking if the pixel was above or below threshold in the SAME image
channel - that's wrong.

Now we check the correct channel, the other one.
As defined by Manders in the original article.
See coloc 2 wiki page.

This bug probably had a small impact as the thresholds of the two
channels tend to be rather similar and quite low.  The results wont
change much in the majority of situations. Only when the thresholds are
very different will be bug be very obvious.

Signed-off-by: daniel j. white <dan@chalkie.org.uk>