Make SpearmanRankCorrelation not static
It was keeping static references to large double arrays,
which made garbage collection impossible after colocalization.

Now it is structured the same as the other algorithms, where you
need to create an instance before executing the methods.

This change breaks backwards API compatibility, but we need to do so,
because those methods never should have been static in the first place.