Main: avoid 'Unrecognized command' error
The recent changes to the codebase somehow broke the main method.
My guess is that the switch to sc.fiji.coloc package prefix is
somehow making ImageJ1 unhappy, but we do set the plugins.dir
appropriately. The plugins.config is there, and correct. So I do
not know why it cannot find the command. To work around the
issue, we use IJ.runPlugIn(className, arg) instead, which works
fine. And then we no longer need to override the plugins.dir.