Fix Colocalisation_Test window crash on multiple runs
If a script uses Colocalisation_Test, and closes the result window,
subsequent attempts to run Colocalisation_Test will crash because the
reference to the window will still be there, but the attempts to write
to the window will crash because the window was closed. To verify,
revert changes to Colocalisation_Test.java & run the new test case added
- the 2nd attempt to run coloc will fail.

The solution was to cleanup the window ref when a window is closed.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>